+++ webpack.config.js
@@ -1,0 +1,24 @@
+const path = require('path');
+
+// TODO: disable sourcemaps in production build mode
+
+module.exports = {
+    entry: './src/js/main.ts',
+    devtool: 'inline-source-map',
+    module: {
+        rules: [
+            {
+                test: /\.tsx?$/,
+                use: 'ts-loader',
+                exclude: /node_modules/,
+            },
+        ],
+    },
+    resolve: {
+        extensions: ['.tsx', '.ts', '.js'],
+    },
+    output: {
+        filename: 'app.bundle.js',
+        path: path.resolve(__dirname, 'src/js'),
+    },
+};